added InputValidator class in services to validate the input given by the user and i have used enum as the return type of InputValidator methods and also improved readability.

+++ ATM.Models/BankAccount.cs
@@ -10,7 +10,7 @@ namespace ATM.Models
     {
         public string userName;
         public string password;
-        public int balance;
+        public double balance;
 
         public BankAccount(string userName, string password)
         {
